Add boilerplate and links
I added the boilerplate to all four pages of the project, as well as the links from the home page to each recipe, and from each recipe to the home page.

The links to the recipes from the home page use "target_blank", but the links from the recipes to the home page do not.

All links use "noopener noreferrer".